pyproject.toml
# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "auto"
log_cli_level = "INFO"

[tool.poetry]
name = "kafka-broker-rack-awareness-operator"
package-mode = false
requires-poetry = ">=2.0.0"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
ops = ">=2.0.0"
pydantic =">=1.10, <2"
pyyaml = "^6.0.1"

[tool.poetry.group.fmt]
optional = true

[tool.poetry.group.fmt.dependencies]
black = "^22.3.0"
ruff = ">=0.0.157"
pyright = "^1.1.300"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^22.3.0"
ruff = ">=0.0.157"
codespell = ">=2.2.2"
pyright = "^1.1.301"

[tool.poetry.group.unit]
optional = true

[tool.poetry.group.unit.dependencies]
pytest = ">=7.2"
coverage = {extras = ["toml"], version = ">7.0"}

[tool.poetry.group.integration]
optional = true

[tool.poetry.group.integration.dependencies]
pytest = ">=7.2"
juju = "^3.6"
coverage = {extras = ["toml"], version = ">7.0"}
pytest-operator = ">0.20"

[tool.poetry.group.format.dependencies]
pyright = "^1.1.301"

# Formatting tools configuration
[tool.black]
line-length = 99
target-version = ["py38"]

# Linting tools configuration
[tool.ruff]
line-length = 99
select = ["E", "W", "F", "C", "N", "D", "I001"]
extend-ignore = [
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
]
ignore = ["E501", "D107"]
extend-exclude = ["__pycache__", "*.egg_info"]
per-file-ignores = {"tests/*" = ["D100","D101","D102","D103","D104"]}

[tool.ruff.mccabe]
max-complexity = 10

[tool.pyright]
include = ["src"]
extraPaths = ["./lib"]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"
reportIncompatibleMethodOverride = false
reportImportCycles = false
reportMissingModuleSource = true
stubPath = ""
